Update license headers on source files to be compliant with LGPL, by using the SPDX identifier.
The project is under LGPL license so the source files should reflect that by having a license header compliant with LGPL, or at least have an appropriate SPDX header.
This changed is required to allow the source files to be used in downstream project because some files are not licensed in accordance to the LPGL rules.
Refs:
- https://www.gnu.org/licenses/gpl-howto.html
- https://spdx.dev/learn/handling-license-info/
- https://spdx.org/licenses/LGPL-3.0-or-later.html